Add 404 handling and align with vike-react conventions
