+++ .gitignore
@@ -22,6 +22,7 @@ npm-debug.log
 node_modules/
 .nyc_output
 tsconfig.tsbuildinfo
+*.tsbuildinfo
 
 ## Project
