--- apps/refresh-22-ui-old/pages/404.tsx
@@ -1,30 +1,0 @@
-import {Header, Paragraph, SubHeader} from "../components/typography";
-
-const NotFound = (): JSX.Element => {
-  const response = (
-    <>
-      <img
-        src={"/art/404.png"}
-        alt={"."}
-        className={"w-96 pt-16 m-auto"}
-      />
-      <Header>
-        404
-      </Header>
-      <SubHeader>
-        Not Found
-      </SubHeader>
-      <Paragraph>
-        Are you lost?
-      </Paragraph>
-    </>
-  );
-
-  return <div className={"flex items-center justify-center h-full"}>
-    <div className={"text-center pb-16"}>
-      {response}
-    </div>
-  </div>;
-};
-
-export default NotFound;
