+++ apps/refresh-22-ui/pages/+config.ts
@@ -1,9 +1,6 @@
 import vikeReact from "vike-react/config";
 
-import Layout from "./Layout.tsx";
-
 export const config = {
-  Layout,
   bodyAttributes: { className: "antialiased" },
   extends: vikeReact,
   ssr: false,
